+++ zoo/src/main/scala/com/intel/analytics/zoo/common/PythonZoo.scala
@@ -15,9 +15,22 @@
  */
 package com.intel.analytics.zoo.common
 
-import com.intel.analytics.bigdl.python.api.{JTensor, PythonBigDLKeras}
+import java.util
+
+import com.intel.analytics.bigdl.python.api.{JTensor, PythonBigDLKeras, Sample}
 import com.intel.analytics.bigdl.tensor.{DenseType, SparseType, Tensor}
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
+import com.intel.analytics.zoo.pipeline.api.Predictable
+import org.apache.spark.api.java.JavaRDD
+import java.util.{List => JList, Map => JMap}
+
+import com.intel.analytics.bigdl.Module
+import com.intel.analytics.bigdl.nn.abstractnn.{AbstractModule, Activity}
+import com.intel.analytics.bigdl.optim.LocalPredictor
+import com.intel.analytics.bigdl.utils.Table
+import com.intel.analytics.zoo.feature.image.ImageSet
+import com.intel.analytics.zoo.feature.text.TextSet
+import scala.collection.JavaConverters._
 
 import scala.reflect.ClassTag
 
@@ -100,5 +113,81 @@ class PythonZoo[T: ClassTag](implicit ev: TensorNumeric[T]) extends PythonBigDLK
           s" ${tensor.getTensorType}")
     }
   }
+
+  def activityToList(outputActivity: Activity): JList[Object] = {
+    if (outputActivity.isInstanceOf[Tensor[T]]) {
+      val list = new util.ArrayList[Object]()
+      list.add(toJTensor(outputActivity.toTensor))
+      list
+    } else {
+      table2JList(outputActivity.toTable)
+    }
+  }
+
+  private def table2JList(t: Table): JList[Object] = {
+    var i = 1
+    val list = new util.ArrayList[Object]()
+    while (i <= t.length()) {
+      val item = t[Object](i)
+      if (item.isInstanceOf[Tensor[T]]) {
+        list.add(toJTensor(item.asInstanceOf[Tensor[T]]))
+      } else if (item.isInstanceOf[Table]) {
+        list.add(table2JList(item.asInstanceOf[Table]))
+      } else {
+        throw new IllegalArgumentException(s"Table contains unrecognizable objects $item")
+      }
+      i += 1
+    }
+    list
+  }
+
+  def zooPredict(
+                  module: Predictable[T],
+                  x: JavaRDD[Sample],
+                  batchPerThread: Int): JavaRDD[JList[Object]] = {
+    module.predict(x.rdd.map(toJSample), batchPerThread).map(activityToList).toJavaRDD()
+  }
+
+  def zooForward(model: AbstractModule[Activity, Activity, T],
+                 input: JList[JTensor],
+                 inputIsTable: Boolean): JList[Object] = {
+    val inputActivity = jTensorsToActivity(input, inputIsTable)
+    val outputActivity = model.forward(inputActivity)
+    activityToList(outputActivity)
+  }
+
+  def zooPredict(
+                  module: Module[T],
+                  x: JList[JTensor],
+                  batchPerThread: Int): JList[JList[Object]] = {
+    val sampleArray = toSampleArray(x.asScala.toList.map{f => toTensor(f)})
+    val localPredictor = LocalPredictor(module,
+      batchPerCore = batchPerThread)
+    val result = localPredictor.predict(sampleArray)
+    result.map(activityToList).toList.asJava
+  }
+
+  def zooPredict(
+                  module: Predictable[T],
+                  x: ImageSet,
+                  batchPerThread: Int): ImageSet = {
+    module.predict(x, batchPerThread)
+  }
+
+  def zooPredict(
+                  module: Predictable[T],
+                  x: TextSet,
+                  batchPerThread: Int): TextSet = {
+    module.predict(x, batchPerThread)
+  }
+
+  def zooPredictClasses(
+                         module: Predictable[T],
+                         x: JavaRDD[Sample],
+                         batchPerThread: Int,
+                         zeroBasedLabel: Boolean = true): JavaRDD[Int] = {
+    module.predictClasses(toJSample(x), batchPerThread, zeroBasedLabel).toJavaRDD()
+  }
+
 
 }
